Update TogglWrapper
過去の履歴から同じ名前のプロジェクトを取得して、設定する
