+++ src/main/java/com/vagivagi/connector/toggl/TogglClient.java
@@ -4,6 +4,7 @@ import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.util.Base64Utils;
 import org.springframework.web.reactive.function.client.WebClient;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 @Component
@@ -18,27 +19,34 @@ public class TogglClient {
                 .build();
     }
 
-    public Mono<TogglResponseBody> getCurrentEntry() {
+    public Mono<TogglTimeEntryResponseBody> getCurrentEntry() {
         return this.webClient.get()
                 .uri("time_entries/current")
                 .retrieve()
-                .bodyToMono(TogglResponseBody.class);
+                .bodyToMono(TogglTimeEntryResponseBody.class);
     }
 
-    public Mono<TogglResponseBody> stopEntry(int timeEntryId) {
+    public Mono<TogglTimeEntryResponseBody> stopEntry(int timeEntryId) {
         return this.webClient.put()
                 .uri("time_entries/" + timeEntryId + "/stop")
                 .retrieve()
-                .bodyToMono(TogglResponseBody.class);
+                .bodyToMono(TogglTimeEntryResponseBody.class);
     }
 
-    public Mono<TogglResponseBody> startEntry(Mono<TogglStartEntryRequestBody> body) {
+    public Mono<TogglTimeEntryResponseBody> startEntry(Mono<TogglStartEntryRequestBody> body) {
         return this.webClient.post()
                 .uri("time_entries/start")
                 .contentType(MediaType.APPLICATION_JSON)
                 .body(body, TogglStartEntryRequestBody.class)
                 .retrieve()
-                .bodyToMono(TogglResponseBody.class);
+                .bodyToMono(TogglTimeEntryResponseBody.class);
+    }
+
+    public Flux<TogglTimeEntry> getTimeEntries() {
+        return this.webClient.get()
+                .uri("time_entries")
+                .retrieve()
+                .bodyToFlux(TogglTimeEntry.class);
     }
 
 }
